refactor: move last_author_command from repo.py to cli.py
