+++ rspec-tools/rspec_tools/cli.py
@@ -14,7 +14,7 @@ from rspec_tools.coverage import (
 )
 from rspec_tools.errors import RuleValidationError
 from rspec_tools.notify_failure_on_slack import notify_slack
-from rspec_tools.repo import last_author_command
+from rspec_tools.repo import get_last_author_for_file
 from rspec_tools.rules import LanguageSpecificRule, RulesRepository
 from rspec_tools.validation.description import (
     validate_parameters,
@@ -171,8 +171,28 @@ def notify_failure_on_slack(message: str, channel: str):
     notify_slack(message, channel)
 
 
-# Add the last-author command to the CLI
-cli.add_command(last_author_command)
+@cli.command("last-author")
+@click.option("--token", envvar="GITHUB_TOKEN", required=True, help="GitHub token")
+@click.option("--user", help="GitHub username")
+@click.option(
+    "--repo",
+    help="Repository in format 'owner/repo'",
+    default=lambda: os.environ.get("GITHUB_REPOSITORY", "SonarSource/rspec"),
+)
+@click.option(
+    "--max-commits", default=3, type=int, help="Maximum number of commits to check"
+)
+@click.argument("file_path")
+def last_author_command(
+    token: str, user: Optional[str], repo: str, max_commits: int, file_path: str
+):
+    """Find the last non-bot GitHub login that modified a given file."""
+    author = get_last_author_for_file(token, repo, file_path, max_commits, user)
+    if author:
+        click.echo(author)
+    else:
+        click.echo("No non-bot author found for the specified file", err=True)
+        exit(1)
 
 
 __all__ = ["cli"]
